Use create date time field function instead of code duplication
